+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## [Unreleased][unreleased]
 
+#### Fixed
+
+* BaseUrl in OpenApiHandler
 
 ### Added